lazily load the FloatingTanStackRouterDevtools
